Added Description
This is a work-in-progress code for the Arduino Uno + ESP8266 + OV7670 combination being used for creating a simple home surveillance system backed by Microsoft Azure IotHub as the data persistence and transmission mechanism. 
The system aims at publishing the data as images/video and receive the same remotely via a Windows (C#) based app on a computer or an Android App for smart phones.